+++ projects/miopen/src/solver/conv_ck_igemm_fwd_bias_res_add_activ_fused.cpp
@@ -74,7 +74,7 @@ bool PerfConfigConvCKIgemmFwdBiasResAddActivFused::SetNextValue(
             return false;
         return true;
     }
-    if((this->index + 1) < valid_kernels.size())
+    if(static_cast<size_t>(this->index + 1) < valid_kernels.size())
     {
         ++this->index;
         this->kernel_id = this->valid_kernels[index];
@@ -86,7 +86,7 @@ bool PerfConfigConvCKIgemmFwdBiasResAddActivFused::SetNextValue(
 
 bool PerfConfigConvCKIgemmFwdBiasResAddActivFused::IsValidValue() const
 {
-    return this->index >= 0 && this->index < valid_kernels.size();
+    return this->index >= 0 && static_cast<size_t>(this->index) < valid_kernels.size();
 }
 
 bool PerfConfigConvCKIgemmFwdBiasResAddActivFused::IsValid(
